Trim namespace from the value passed to print-doc
Fixes an issue in which the namespace would appear twice when calling doc. If
the namespace on the :name symbol exists and is the same as the :ns symbol,
we drop the namespace value from the :name symbol before calling print-doc.

Related to CLJS-921, but the fix should not cause problems if ClojureScript's
behavior changes.