Unquoting and tests for with-open
Because of #204 we need to unquote finally

We also cannot let syntax quoting work on -close, otherwise
it will end up fully qualifying as planck.core$macros/-close.

Add tests. These have to be integration tests done with the
binary because of the way the core.clj file is loaded and because
we need macroexpansion to be tested via bootstrap. (In other
words, it can't be done via the unit tests.)