If not develping, prefer loading bundled source
When you are developing Planck itself, it is useful to be able to, from
withing the Planck REPL, load and reload the ClojureScript source
implementing Planck.

But, in production, this isn't needed—you want to load bundled code in
preference to all other code. This is especially true if ClojureScript
or other similar JARs are included on the classpath—you don't want to
inadvertently load that incompatible code.

Fixes #273