Start port of Planck to C
This is the start of a port of Planck to C.  Right now it has support
for the REPL (with syntax highlighting, basic line editing and history),
init and main opts, but not much more.

To build it, run `./script/build-c` from the top-level directory.

The code should support Linux and OSX, but was only tested on Linux.
BSDs should also be fairly easy to support in the future.

Things that are missing:

- anything requiring async and/or thread support
    (can be added using pthreads)

    This also means that there is currently no support for
    initialization in a thread, socket REPLs, ...
- encoding handling (so far, input is assumed to be UTF8)
- a lot of the exported js/PLANCK_* functions (see below)
- fancier REPL features (paren matching, live highlighting, ...)
- and more that I don't know about

The following js/PLANCK_* functions are still missing, although some of
them are fairly easy to add:

- js/PLANCK_DELETE
- js/PLANCK_FILE_INPUT_STREAM_CLOSE
- js/PLANCK_FILE_INPUT_STREAM_OPEN
- js/PLANCK_FILE_INPUT_STREAM_READ
- js/PLANCK_FILE_OUTPUT_STREAM_CLOSE
- js/PLANCK_FILE_OUTPUT_STREAM_OPEN
- js/PLANCK_FILE_OUTPUT_STREAM_WRITE
- js/PLANCK_FILE_READER_CLOSE
- js/PLANCK_FILE_READER_OPEN
- js/PLANCK_FILE_READER_READ
- js/PLANCK_FILE_WRITER_CLOSE
- js/PLANCK_FILE_WRITER_OPEN
- js/PLANCK_FILE_WRITER_WRITE
- js/PLANCK_FSTAT
- js/PLANCK_GET_EXIT_VALUE
- js/PLANCK_IS_DIRECTORY
- js/PLANCK_LIST_FILES
- js/PLANCK_READ_PASSWORD
- js/PLANCK_REQUEST
- js/PLANCK_SET_EXIT_VALUE
- js/PLANCK_VERSION