Initialize ClojureScript in a background thread (#314)
* Free temporary string when resolving namespaces

* Restructure initialization to happen outside of main()

This means the configuration is now a global struct, that the functions
exposed are defined elsewhere, and that initialization now happens in
cljs_engine_init.

* Initialize cljs in a background thread

This works, but I'm not sure why we have to block SIGUSR2.  Otherwise
we don't accept input anymore, and the process seems to hang.  I don't
know why we get a SIGUSR2, and I only found out because I was debugging
in gdb.

* Fix some small problems in main()

* Add comment about the SIGUSR2 mystery