Replace all references to 'accepts' with 'accept' (#862)
* Replace all references to 'accepts' with 'accept'

The `planck.http` library allows a user to set the ACCEPT header in HTTP
requests. However, the library currently refers to the option to set this
header, as well as related functions, using the word 'accepts' rather than
'accept' (see #837).

This PR replaces all references to 'accepts' with 'accept'. It is arguable
whether these changes represent a breaking change. Planck's documentation
previously said that `planck.http/get` (and by extension `patch`, `post` and
`put`) could be passed an `opts` map with an `:accepts` key. However, the
private functions within `planck.http/get` always used `:accept` as the key so
one could view this as a bug fix rather than a change to the API.

The tests and documentation for `plank.http/get` have also been updated.

* Update changelog