Document using Boot to generate a classpath file
Planck can access dependencies that are specified on its classpath.
These can be specified by hand but an alternative is to use a tool like
Boot to output the dependencies to a file that can then be fed into
Planck. This PR adds documentation to the Planck docs that explains how
this can be done and closes #343.

It's noted that, due to the introduction of tools.deps and Planck's
support for it via `plk`, this is less important than it was at the
time of the initial discussion.