Support interning quoted data structures (#508)
Planck provides an `intern` function in `planck.core` that is similar
to Clojure's `clojure.core/intern`. Whereas Clojure's version calls out
to a Java class, Planck `eval`s a `def` statement defining the name and
value to be interned.

The prior implementation in Planck used syntax quoting to generate the
`def` form. The problem with this was that certain data structures that
would be passed to `planck.core/intern` as a quoted argument would be
unquoted in the `def` form. Depending on the data structure, this could
then cause an error.

A simple example that demonstrates the problem is passing a symbol as
the value to be bound. Calling `(planck.core/intern *ns* 'foo 'bar)`
would result in one of two things, both of which were wrong. If 'bar`
were not bound, Planck would report an error when `bar` was unquoted
because no value existed. If `bar` were bound, that value would be
interpolated. In neither case would `foo` be bound to the symbol `bar`.

This commit adds support for quoted data structures by quoting any value
that is passed to `planck.core/intern`. This takes advantage of the fact
that quoting has no effect on most data structures. For example, `'5`
results in `5`. This closes #508.