Add almost all of the Google Closure Library (#496)
Due to the way Planck is built, the Google Closure Library needs to be
bundled together with the executable. While the entirety of the library
has historically been available during the build process, only a small
subset has actually been included in the bundle.

This commit broadens the scope to almost the entirety of the Google
Closure Library (save for the facilities made available under the
`goog.demos` namespace). While this provides broader support to users,
it does substantially increase the amount of time it takes to complete
the full build process.

To provide an up to date version of the Google Closure Library, this
commit switches the version that is downloaded from the Maven repository
to a more recent distribution. This newer version of the Google Closure
Library caused conformance warnings to be displayed during the build
process. To avoid this, a newer version of the Google Closure Compiler
is also now required. As the newer version of the Google Closure Library
relies on features from ECMAScript 6, the language output of the Closure
Compiler has been switched to this version of ECMAScript.

Finally, the documentation has been updated as required.