Remove some bundled GCL namespaces (#496)
The initial commit included almost the entirety of the Google Closure
Library. Doing so increased the time taken to compile Planck
considerably. This commit removes some of the namespaces that are not
considered appropriate or desirable for Planck. This reduces the time
taken for a full build by about a third.

The documentation has also been updated.