Restore GCL localisations
To improve build times for the optimised version of Planck,the previous
commit did not include the localisation namespaces that are part of the
Google Closure Library. This commit restores the namespaces and instead
changes the `planck-cljs/script/bundle` script to skip these files when
optimising.

This commit can now be considered to fix #496. The documentation and
change log have been updated accordingly.