Improve Google Closure Library Bundling (#901)
Planck bundles the Google Closure Library. Previous to this commit, that
was achieved by: (1) installing an unofficial distribution of GCL from
Maven Central; and (2) manually listing the namespaces to import in
`planck-cljs/bundle.cljs`. This was problematic since it meant that the
GCL bundled with Planck was often not up to date. It also was difficult
to maintain as the large number of namespaces made available via GCL
were hand-coded.

This commit solves those problems by having the build script: (1)
download an official distribution of GCL from the official GitHub
repository; and (2) generate the namespaces to import from the files
that make up this distribution.

In addition, this commit also introduces a caching mechanism for the
files generated as part of the build process. Planck provides a `--fast`
compilation mode that skips some of the optimisation steps that greatly
increase build times but even this mode repeats work that is unnecessary
for the majority of the bundled files that do not change.

The cache works by saving the generated files to the cache with a name
that combines the name of the file and the MD5 hash of the file's
contents.