Fix single-character single-segment namespace crash (#900)
The Planck REPL provides for multiline input with automatic indentation
for subsequent lines. This is achieved using the Linenoise library. When
defining a prompt using Linenoise, a secondary prompt is provided to
display on subsequent lines. Planck's secondary prompt is "#_=>". This
results in an expected minimum width for any Clojure namespace being
displayed at the REPL as two characters.

The `form_prompt()` function in `repl.c` contained an arithmetic
operation that resulted in a negative value if the first operand
(representing the length of the namespace) was less than 2 (as discussed
in #588). This commit tests for that condition and provides that the
result cannot be less than zero. It also renames some of the variables
used for clarity.

This fixes #588.