Demonstrate pREPL proof of concept
This commmit demonstrates a proof of concept pREPL. This pREPL operates
differently to Clojure's pREPL (and ClojureScript's pREPL) in the
following ways:

1. tightly coupled with existing REPL;
2. looping logic is written in C;
3. different output formats; and
4. no tap functions.

Tightly Coupled with Existing REPL
----------------------------------

As explained in more detail below, Planck implements its REPL through a
mixture of C and Clojure code. The looping function is written in C and
the evaluation and processing functions are largely written in
ClojureScript.

Because of this, the ClojureScript code does not include the same
utility functions included in Clojure's REPL for setting up the
evaluation and processing environment. Rather, these values are set
within the evaluation and processing functions themselves.

The tighter coupling of setup and evaluation in Planck's ClojureScript
code makes it difficult to implement the demonstrated pREPL in as
separate a manner as Clojure's pREPL.

While it is theoretically possible to refactor Planck's existing REPL
code to allow this, for the purposes of the proof of concept, it was
decided to piggyback on the existing REPL evaluation and processing
functions. This required minor changes, mostly the introduction of
default output functions that could be overridden by the pREPL.

Looping Logic Written in C
--------------------------

The fact that the demonstrated pREPL piggybacks on the existing REPL
code means that the implementation works as the same mixture of C and
ClojureScript, with the looping logic being written in C.

Since the loop is written in C, it was more difficult to implement a
general pREPL that could have its input and output values set in a
calling function. An IO pREPL similar to the one provided in Clojure and
ClojureScript has been implemented instead.

Relatedly, it was also more difficult to provide a means for the user to
designate a specific function to perform as the pREPL. Clojure and
ClojureScript require the user to specify the specific pREPL
implementation. The demonstrated pREPL includes a new command line
option for setting up the pREPL with a given host and port (patterned on
the existing socket REPL option).

Different Output Formats
------------------------

pREPL functionality remains in an alpha state and the implementations in
Clojure and ClojureScript are not consistent in the way that they output
values. Moreover, the pREPL provided by ClojureScript is itself written
in JVM Clojure. As such, there is no canonical ClojureScript
implementation. The demonstrated pREPL aims to be consistent with the
Clojure pREPL.

No Tap Functions
----------------

The Clojure and ClojureScript pREPLs include a tap function that can be
used within the loop to output data in an EDN data structure that is
separate to the return, printing and error data structures. This tap
function is created before the loop begins. Since that is not possible
using the current architecture, the demonstrated pREPL does not provide
any tap functions.