Better separate pREPL code
This version of the pREPL proof of concept better separates the pREPL
code from the existing REPL code.

The pREPL C code is split out into its own source and header files. The
pREPL loop now sets up and tears down the tap function and avoids having
to rebind the *exec-tap-fn* to transmit data to the socket on which the
pREPL is connected.

The ClojureScript code moves the input, output and tap functions
(conceptually thought of as 'channels') into a single function that can
be, and indeed is, called by a setup function. This moves the code
closer to the reference implementation.