Fix -e arguments affecting *1, *2, *3 and *e
Planck's evaluation logic treated evaluation of text entered at a REPL
and evaluation of arguments to `-e` the same. This resulted in the
result variables (`*1`, `*2`, `*3`) being set in both cases. As noted
in #659, this is not the behaviour of Clojure. This commit fixes #659 by
testing whether the evaluation is occurring in a REPL environment.