Fix empty lines not being processed
As described in #661, Planck has a bug where an empty line after a
backslash is not processed and so does not produce the correct output (a
newline character). This was due to the fact a newline by itself appears
to the REPL as an empty string. This commit detects that and processes
it properly. This fixes #661.